% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hydrophobicity.R
\name{calc_hydrophobicity}
\alias{calc_hydrophobicity}
\title{Calculate Gravy (Hydrophobicity) Index of a Protein Sequence}
\usage{
calc_hydrophobicity(sequence)
}
\arguments{
\item{sequence}{Karakter vektörü. Tek harfli amino asit dizisi (örn: "MKVLWA").}
}
\value{
Sayısal bir değer (GRAVY skoru). Pozitif değerler hidrofobik,
negatif değerler hidrofilik karakteri gösterir.
}
\description{
Bu fonksiyon verilen bir amino asit dizisinin Kyte-Doolittle hidrofobiklik
indeksine göre ortalama hidrofobikliğini (GRAVY skoru) hesaplar.
}
\examples{
calc_hydrophobicity("MKVLWA")
}
