% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mw.R
\name{calc_mw}
\alias{calc_mw}
\title{Calculate Molecular Weight of a Protein Sequence}
\usage{
calc_mw(sequence)
}
\arguments{
\item{sequence}{A character string representing the protein sequence.}
}
\value{
A numeric value representing the molecular weight in Daltons.
}
\description{
Calculate Molecular Weight of a Protein Sequence
}
