% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-calc_mw.R
\name{calc_mw}
\alias{calc_mw}
\title{Calculate Molecular Weight of a Protein Sequence}
\usage{
calc_mw(sequence)
}
\arguments{
\item{sequence}{A character string representing the protein sequence (single-letter codes).}
}
\value{
A numeric value representing the molecular weight in Daltons (Da).
}
\description{
This function calculates the molecular weight (in Daltons) of a given
amino acid sequence based on average isotopic weights, taking into account
the loss of water molecules during peptide bond formation.
}
\examples{
calc_mw("ACDEFGHIKLMNPQRSTVWY")
}
